refactor helper | fix blockly logic
