+++ blockly/src/client/Client.java
@@ -14,7 +14,7 @@ import core.systems.LevelSystem;
 import core.systems.PlayerSystem;
 import core.utils.Tuple;
 import core.utils.components.path.SimpleIPath;
-import entities.utility.HeroTankController;
+import entities.HeroTankControlledFactory;
 import java.io.IOException;
 import java.util.logging.Level;
 import level.MazeLevel;
@@ -116,7 +116,7 @@ public class Client {
   private static void createHero() {
     Entity hero;
     try {
-      hero = HeroTankController.newTankControlledHero();
+      hero = HeroTankControlledFactory.newTankControlledHero();
     } catch (IOException e) {
       throw new RuntimeException(e);
     }
